Renamed Folder and removed GBRT Surrogate Model
