+++ app/create/bo/sbo_components/chooseAcquisition.js
