+++ app/create/bo/sbo_components/chooseAlgorithmType.js
