+++ app/create/bo/sbo_components/configureBounds.js
