+++ app/create/bo/sbo_components/configureInitialDesign.js
