+++ app/create/bo/sbo_components/configureKernel.js
